requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
gunicorn>=21.2
pydantic>=2.6
pydantic-settings>=2.2
motor>=3.3
pymongo>=4.6
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
email-validator>=2.1
python-multipart>=0.0.9